refactor(perps): extract payment token init logic into useInitPerpsPaymentToken hook
Move payment token initialization (pending config restoration, default
token selection, cleanup) from PerpsPayRow and PerpsOrderView into a
dedicated hook, fixing missing imports and variable declarations.
